retire in-house pubsub
we don't need a pubsub to subscribe only to the first published valuee, we can use  an expectant store that returns a promise when the store does not yet contain the value, where all the relevant promises resolve on a future set.

this change means we can retire the in-house pubsub.

repeaters probably make in-house pubsub unnecessary anyway, and were not yet used elsewhere in the codebase

also somewhat relevant, see https://github.com/repeaterjs/repeater/issues/67#issue-692118936